리팩토링
- as-is
   - index.js 파일 하나에 모두 작성
- to-be
   - index.js에서 클래스별로 리팩토링
   - Board.js
   - Game.js
   - Square.js
