+++ my-app/src/Square.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+export class Square extends React.Component {
+    render() {
+      return (
+        <button 
+        className="square" 
+        onClick={() => this.props.onClick()}
+        >
+        {this.props.value}
+        </button>
+      );
+    }
+}
